+++ enclave_apps/key_xor_test_app/src/main.rs
@@ -21,7 +21,10 @@
 extern crate alloc;
 
 use oak_restricted_kernel_interface::{syscall::read, DERIVED_KEY_FD};
-use oak_restricted_kernel_sdk::{entrypoint, FileDescriptorChannel, Read, Write};
+use oak_restricted_kernel_sdk::{
+    channel::{FileDescriptorChannel, Read, Write},
+    entrypoint,
+};
 
 // Continuously reads single bytes from the communication channel, XORs them with a byte from the
 // derived key and sends them back.
